feat: updated song deletion implementation

+++ src/domain/useCases/RemoveSongUseCase.ts
@@ -2,6 +2,7 @@ import { resolve } from "path";
 import EmitterManager from "../../common/EmitterManager";
 import { Event } from "../../common/Event";
 import SongRepository from "../../data/repositories/SongRepository";
+import ReservedSongRepository from "../../data/repositories/ReservedSongRepository";
 
 export default interface RemoveSongUseCase {
   execute(id: string): Promise<void>;
@@ -9,13 +10,21 @@ export default interface RemoveSongUseCase {
 
 export class DefaultRemoveSongUseCase {
   songRepository: SongRepository;
+  reservedSongRepository: ReservedSongRepository;
 
-  constructor(songRepository: SongRepository) {
+  constructor(songRepository: SongRepository, reservedSongRepository: ReservedSongRepository) {
     this.songRepository = songRepository;
+    this.reservedSongRepository = reservedSongRepository;
   }
 
   async execute(id: string): Promise<void> {
-    console.log(`Removing reserved song with id: ${id}`);
+    // first, check if the song is reserved
+    const reservedSongs = await this.reservedSongRepository.getReservedSongs();
+    const reservedSong = reservedSongs.find((song) => song.song.identifier === id);
+    if (reservedSong) {
+      throw new Error(`Song with id: ${id} is reserved`);
+    }
+    console.log(`Removing song with id: ${id}`);
     await this.songRepository.deleteSongRecord(id);
   }
 }
